fix paths after moving folders
